+++ tests/x509/test_x509.py
@@ -1232,18 +1232,20 @@ class TestRSACertificateRequest(object):
         assert isinstance(extensions, x509.Extensions)
         assert list(extensions) == []
 
-    def test_challenge_password(self, backend):
+    def test_get_attribute_for_oid(self, backend):
         request = _load_cert(
             os.path.join(
                 "x509", "requests", "challenge.pem"
             ), x509.load_pem_x509_csr, backend
         )
-        assert request.challenge_password == b"challenge me!"
+        assert request.get_attribute_for_oid(
+            x509.oid.AttributeOID.CHALLENGE_PASSWORD
+        ) == b"challenge me!"
 
-    def test_invalid_challenge_password(self, backend):
+    def test_invalid_attribute_for_oid(self, backend):
         """
         This test deliberately triggers an InternalError because to parse
-        challenge passwords we need to do a C cast. If we're wrong about the
+        CSR attributes we need to do a C cast. If we're wrong about the
         type that would be Very Bad so this test confirms we properly explode
         in the presence of the wrong types.
         """
@@ -1253,7 +1255,9 @@ class TestRSACertificateRequest(object):
             ), x509.load_der_x509_csr, backend
         )
         with pytest.raises(InternalError):
-            request.challenge_password
+            request.get_attribute_for_oid(
+                x509.oid.AttributeOID.CHALLENGE_PASSWORD
+            )
 
     def test_no_challenge_password(self, backend):
         request = _load_cert(
@@ -1261,7 +1265,11 @@ class TestRSACertificateRequest(object):
                 "x509", "requests", "rsa_sha256.pem"
             ), x509.load_pem_x509_csr, backend
         )
-        assert request.challenge_password is None
+        with pytest.raises(x509.AttributeNotFound) as exc:
+            request.get_attribute_for_oid(
+                x509.oid.AttributeOID.CHALLENGE_PASSWORD
+            )
+        assert exc.value.oid == x509.oid.AttributeOID.CHALLENGE_PASSWORD
 
     @pytest.mark.parametrize(
         "loader_func",